Automatically include digitalWriteFast
I included it from Arduino.h

Note: there is a work around included in the digitalWriteFast, that it will only include itself if PINS_COUNT is defined.

PINS_COUNT is defined at the end of variant.h.  So this makes sure that the full variant.h has been included.  As there is an issue that if variant.h is included first, before Arduino.h both of which include each other,  the include of the digitalWriteFast.h will try to be processed before the pin data is defined and the compiler will error out.