Wire Library: fix endTransfer where sendStop=0
Fix endTransfer/requestFrom: to work after a call to either of the two functions when senStop=0.

Issue was in these cases the buss will show as busy which we would then have aborted the call.  Now remembers if it's last call said to not send the Stop condition and then continue in this case.  At which point Wire will send a restart condition.

This fixed issue with MLX90614 Temp Sensor