Asynch Serial more minimal like changes
Changed to use a more minimal change to underlying system.

Uses a modified version of the current built in TX Buffers and depending on
constructor option uses the HAL_UART_Transmit_IT function, or the
HAL_UART_Transmit_DMA function.

Note: There is an issue with using DMA on buffers that don't have the
attribute attribute((section(".NoneCacheableMem")))

So I changed the TX buffer to have the actual buffer with this
attribute, I defined the buffer in the variant.cpp and then passed in
the buffer as well as the size to the constructor of the Serial class.
This should allow us to tailer the actual sizes needed depending on
which Serial port we use.

Also added the SerialX.write(buf, cnt) implementation, which is used for example in PortHandler write functions.  What this code does, is to move as much as it can into the ouptut buffer, and then call the HAL transfer on as many bytes as it can.  If any more bytes are written while that one is busy, it continues to put into the output buffer.  When the first one completes.  On the UARTClass:TxHandler function sees if any additional stuff is in the buffer, if so again it issues a request for as many bytes as it can.

Note: This implemention has same issue as USB one, that if the writes wrap around in the buffer, it only writes up to the end of the buffer in one output and then calls again starting at start of buffer.

I also have code in place to allow me to have the UArtClass code control the HalfDuplex switching (Turns on when writes happen, turns off if TXHandler is called and no more bytes to output.

Warning  Serial3 uses DMA.  So trying to work it out.

Note: this also breaks some functionality on SPI2 ...

Should hopefully fix in next version