Workbench Compiler warnings
Did not want to look at the warnings any longer as I turn on show verbose compile and show all warnings.

Some of these changes may not necessary, but some maybe stop faults.  Like trying to do a  sync read or write to an invalid item name, will not find it in list, but then use swr or swh or srh data which was uninitialized.