Dynamixel Workbench:  Update to version OpenCM/Main Workbench project
The library had the first set of changes before I updated OpenCM to share the names of items, and rearrange items, so that you could compare and decide which way you wanted to go.

As you choose the more complete change (which I prefered), here is the updates to make OpenCR the same as the others.

Note: I had one extra change in line for Uart Class that defined SERIAL_WRITES_NON_BLOCKING, so if we updated Dynamixel SDK to include new stuff, it could check to know the user is running a version of OpenCR libraries that support non-blocking writes and you need to do flushes...