Dynamixel SDK: Sync Read/Write two call init option to allow global objects
Currently if you global Sync Read or write as a global object, it will crash the processor, as you can not rely on when the constructor is called.  Also the passed in port and protocol pointers are not set yet.

So created a version of the constuctor that you don't pass these two pointers in, but instead call init(port, ph), to fill it in during setup.

My test case works with this :D