Dynamixel SDK:group_sync_read - reduce/eliminate memory allocations
Like the prevous update, changed group_sync_read to work with just one allocated buffer.  Calling program also has the option to set this buffer as to be able to statically define memory usages.

Some updates to the Sync_write code in previous commit to move all of the memory allocation and adding items to the look function.

Also have a value defined of how much to add memory the user needs to add for each servo item to the count of bytes returned or set in the Sync read or write.

Again test case updated