Update arduino/opencr_arduino/opencr/libraries/OpenCR/examples/08. DynamixelWorkbench/t_Find_Dynamixel/t_Find_Dynamixel.ino
Workbench: Find Dynamixels test make more Arduino like.

Move code to loop, which allows you to rerun the test, which I like as sometimes I forget to turn on the Servo power and the servos are not found.

Also changed the printing of servo info to be more Arduino.  That is simply use simple Serial.print functions instead of generating and combing string.