Turtlebot 3 setup motors - Maybe a little more robust
Try to make the code a little more robust to handle when the servos are somewhat setup.

So in different places the code knows which servo id it is looking for and if found it tries to use that one.  Where before if you had servo ID #1 and #2 and was trying to setup left, it would grab the last servo found #2 and then set it's id to #1...

Also used ping to know if the specific servo was found that is expected for the test servo cases and only does this at baud of 1000000 as that is what turtle expects